revert to built-in sdk/widget for < 29
toolbar widget gives users a consistent experience across FF < 29 and >= 29.
After releasing the alpha, it became clear that this was not worthwhile. Users
of < 29 expect addons to appear in the addon bar (and they can always be
dragged to the toolbar, which works well). Additionally, using toolbarwidget
created a problem where moving the icon from the toolbar to the addon bar did
not lead to the icon being properly resized.