Fixes to Swedish locale
Made some improvements to the Swedish translation. This is only improved from the previous translation and not based off the original in English.

What I observed was the usage of "tracka" which is used as short form in everyday language for "trakassera" (English: "to harass"). Used the word "spåra" instead, which is "to track". "Tracker" is however the used technical term (to the best of my knowledge) but I swapped it for "spanare", which would be the proper translation.

Found a bunch of composition errors too and some improvements to the choices of words.