Remove the reset button if it's not required
The reset to default button is removed in the popup
and options if it's not required (if the user didn't
override anything for this origin). Before this commit
a blank popup is appearing if hovering the empty spot
which a) is annoying and b) is decreasing HTML parsing
and DOM performance a bit.